# lib/api-spec/openapi_half.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PMFBY Crop Insurance AI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /claims:
    get:
      operationId: listClaims
      tags: [claims]
      summary: List insurance claims
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, under_review, approved, rejected, flagged]
        - name: district
          in: query
          schema: { type: string }
        - name: cropType
          in: query
          schema: { type: string }
        - name: damageType
          in: query
          schema: { type: string }
        - name: fromDate
          in: query
          schema: { type: string, format: "date" }
        - name: toDate
          in: query
          schema: { type: string, format: "date" }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Claim"
    post:
      operationId: createClaim
      tags: [claims]
      summary: Submit a new insurance claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimInput"
      responses:
        "201":
          description: Claim created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
  /claims/{id}:
    get:
      operationId: getClaim
      tags: [claims]
      summary: Get claim details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Claim details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "404":
          description: Not found
    patch:
      operationId: updateClaim
      tags: [claims]
      summary: Update claim status or details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimUpdate"
      responses:
        "200":
          description: Updated claim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
  /claims/{id}/approve:
    post:
      operationId: approveClaim
      tags: [claims]
      summary: Approve a claim
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimDecision"
      responses:
        "200":
          description: Claim approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
  /claims/{id}/reject:
    post:
      operationId: rejectClaim
      tags: [claims]
      summary: Reject a claim
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimDecision"
      responses:
        "200":
          description: Claim rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
  /claims/{id}/flag:
    post:
      operationId: flagClaim
      tags: [claims]
      summary: Flag a claim for manual review
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimFlag"
      responses:
        "200":
          description: Claim flagged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"

  # ─── AI Predictions ───────────────────────────────────────────────────────
...
  /anomalies:
    get:
      operationId: listAnomalies
      tags: [anomalies]
      summary: List detected anomalies and fraud flags
      parameters:
        - name: claimId
          in: query
          schema: { type: integer }
        - name: anomalyType
          in: query
          schema: { type: string }
        - name: resolved
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of anomalies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Anomaly"
  /anomalies/{id}:
    get:
      operationId: getAnomaly
      tags: [anomalies]
      summary: Get anomaly details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Anomaly details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anomaly"
        "404":
          description: Not found
  /anomalies/{id}/resolve:
    post:
      operationId: resolveAnomaly
      tags: [anomalies]
      summary: Mark anomaly as resolved
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnomalyResolution"
      responses:
        "200":
          description: Anomaly resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anomaly"

  # ─── Weather ─────────────────────────────────────────────────────────────
...
  /data-collection:
    get:
      operationId: listDataCollections
      tags: [datacollection]
      summary: List field data collection records for annotation
      parameters:
        - name: cropType
          in: query
          schema: { type: string }
        - name: annotated
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of data collection records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataCollection"
    post:
      operationId: createDataCollection
      tags: [datacollection]
      summary: Submit a field data collection record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataCollectionInput"
      responses:
        "201":
          description: Data collection record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataCollection"
  /data-collection/{id}:
    get:
      operationId: getDataCollection
      tags: [datacollection]
      summary: Get data collection record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Data collection record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataCollection"
        "404":
          description: Not found
    patch:
      operationId: updateDataCollection
      tags: [datacollection]
      summary: Update annotation for a data collection record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataCollectionAnnotation"
      responses:
        "200":
          description: Annotated record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataCollection"
...